requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-cors-headers==3.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Pillow==8.3.2
pycodestyle==2.7.0
PyJWT==2.3.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
